+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/HubdleProperty.kt
@@ -19,11 +19,16 @@ public object HubdleProperty {
     }
 
     public object Analysis {
-        public const val hostUrl: String = "analysis.sonar.host.url"
-        public const val login: String = "analysis.sonar.login"
-        public const val organization: String = "analysis.sonar.organization"
-        public const val projectKey: String = "analysis.sonar.projectKey"
-        public const val projectName: String = "analysis.sonar.projectName"
+        public object Qodana {
+            public const val projectKey: String = "analysis.qodana.projectKey"
+        }
+        public object Sonar {
+            public const val hostUrl: String = "analysis.sonar.host.url"
+            public const val login: String = "analysis.sonar.login"
+            public const val organization: String = "analysis.sonar.organization"
+            public const val projectKey: String = "analysis.sonar.projectKey"
+            public const val projectName: String = "analysis.sonar.projectName"
+        }
     }
 
     public object JetBrains {