+++ CHANGELOG.md
@@ -14,6 +14,12 @@
 
 ### Updated
 
+- `io.kotest:kotest-runner-junit5 -> 5.5.5`
+- `io.kotest:kotest-runner-junit4 -> 5.5.5`
+- `io.kotest:kotest-property -> 5.5.5`
+- `io.kotest:kotest-assertions-sql -> 5.5.5`
+- `io.kotest:kotest-assertions-json -> 5.5.5`
+- `io.kotest:kotest-assertions-core -> 5.5.5`
 - `org.jetbrains.intellij.plugins:gradle-intellij-plugin -> 1.13.2`
 - `app.cash.molecule:molecule-gradle-plugin -> 0.8.0`